+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/content/advancements/criterion/BreedCowWithTypeTrigger.java
@@ -24,20 +24,22 @@ public class BreedCowWithTypeTrigger extends SimpleCriterionTrigger<BreedCowWith
     public static final BreedCowWithTypeTrigger INSTANCE = new BreedCowWithTypeTrigger();
     public static final ResourceLocation ID = BovinesAndButtercups.asResource("breed_cow_with_type");
     public static final Codec<TriggerInstance> CODEC = RecordCodecBuilder.create(inst -> inst.group(
+            ContextAwarePredicate.CODEC.optionalFieldOf("player").forGetter(BreedCowWithTypeTrigger.TriggerInstance::player),
             CowTypeType.CODEC.optionalFieldOf("cow_type_type").forGetter(TriggerInstance::typeType),
             RegistryCodecs.homogeneousList(BovinesRegistryKeys.COW_TYPE).optionalFieldOf("cow_types", HolderSet.direct()).forGetter(TriggerInstance::types),
             ContextAwarePredicate.CODEC.optionalFieldOf("parent").forGetter(TriggerInstance::parent),
             ContextAwarePredicate.CODEC.optionalFieldOf("partner").forGetter(TriggerInstance::partner),
-            ContextAwarePredicate.CODEC.optionalFieldOf("child").forGetter(TriggerInstance::child)
+            ContextAwarePredicate.CODEC.optionalFieldOf("child").forGetter(TriggerInstance::child),
+            Codec.BOOL.optionalFieldOf("different_from_parents").forGetter(TriggerInstance::differentFromParents)
     ).apply(inst, TriggerInstance::new));
 
     private BreedCowWithTypeTrigger() {}
 
-    public void trigger(ServerPlayer serverPlayer, Animal parent, Animal partner, AgeableMob child, Holder<CowType<?>> moobloomType) {
+    public void trigger(ServerPlayer serverPlayer, Animal parent, Animal partner, AgeableMob child, boolean differentFromParents, Holder<CowType<?>> type) {
         LootContext parentContext = EntityPredicate.createContext(serverPlayer, parent);
         LootContext partnerContext = EntityPredicate.createContext(serverPlayer, partner);
         LootContext childContext = child != null ? EntityPredicate.createContext(serverPlayer, child) : null;
-        trigger(serverPlayer, (triggerInstance) -> triggerInstance.matches(parentContext, partnerContext, childContext, moobloomType));
+        trigger(serverPlayer, (triggerInstance) -> triggerInstance.matches(parentContext, partnerContext, childContext, differentFromParents, type));
     }
 
     @Override
@@ -45,19 +47,21 @@ public class BreedCowWithTypeTrigger extends SimpleCriterionTrigger<BreedCowWith
         return CODEC;
     }
 
-    public record TriggerInstance(Optional<Holder<CowTypeType<?>>> typeType,
+    public record TriggerInstance(Optional<ContextAwarePredicate> player,
+                                  Optional<Holder<CowTypeType<?>>> typeType,
                                   HolderSet<CowType<?>> types,
                                   Optional<ContextAwarePredicate> parent,
                                   Optional<ContextAwarePredicate> partner,
-                                  Optional<ContextAwarePredicate> child) implements SimpleInstance {
+                                  Optional<ContextAwarePredicate> child,
+                                  Optional<Boolean> differentFromParents) implements SimpleInstance {
 
-        public boolean matches(LootContext parentContext, LootContext partnerContext, LootContext childContext, Holder<CowType<?>> type) {
-            return type.isBound() && (typeType.isEmpty() && types.size() == 0 || typeType.isPresent() && typeType.get() == type.value().type() && types.contains(type)) && (this.child.isEmpty() || this.child.get().matches(childContext)) && (this.parent.isEmpty() || this.parent.get().matches(parentContext)) && (this.partner.isEmpty() || this.partner.get().matches(partnerContext)) || (this.parent.isEmpty() || this.parent.get().matches(partnerContext)) && (this.partner.isEmpty() || this.partner.get().matches(parentContext));
+        public boolean matches(LootContext parentContext, LootContext partnerContext, LootContext childContext, boolean isNewType, Holder<CowType<?>> type) {
+            return type.isBound() && (typeType.isEmpty() || typeType.get().isBound() && typeType.get().value() == type.value().type()) && (types.size() == 0 || types.contains(type)) && (this.child.isEmpty() || this.child.get().matches(childContext)) && ((this.parent.isEmpty() || this.parent.get().matches(parentContext)) && (this.partner.isEmpty() || this.partner.get().matches(partnerContext)) || (this.parent.isEmpty() || this.parent.get().matches(partnerContext)) && (this.partner.isEmpty() || this.partner.get().matches(parentContext))) && (differentFromParents.isEmpty() || isNewType == differentFromParents.get());
         }
 
         @Override
         public Optional<ContextAwarePredicate> player() {
-            return Optional.empty();
+            return player;
         }
     }
 }