+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/mixin/PistonStructureResolverMixin.java
@@ -1,6 +1,7 @@
 package house.greenhouse.bovinesandbuttercups.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import house.greenhouse.bovinesandbuttercups.api.BovinesTags;
 import house.greenhouse.bovinesandbuttercups.registry.BovinesBlocks;
 import net.minecraft.world.level.block.piston.PistonStructureResolver;
 import net.minecraft.world.level.block.state.BlockState;
@@ -16,7 +17,7 @@ public class PistonStructureResolverMixin {
 
     @ModifyReturnValue(method = "canStickToEachOther", at = @At("RETURN"))
     private static boolean bovinesandbuttercups$dontMakeRichHoneyStickToOtherPistonBlocks(boolean original, BlockState state, BlockState state2) {
-        if (state.is(BovinesBlocks.RICH_HONEY_BLOCK) && !state2.is(BovinesBlocks.RICH_HONEY_BLOCK) || !state.is(BovinesBlocks.RICH_HONEY_BLOCK) && state2.is(BovinesBlocks.RICH_HONEY_BLOCK))
+        if (state.is(BovinesBlocks.RICH_HONEY_BLOCK) && !state2.is(BovinesTags.BlockTags.DOES_NOT_STICK_RICH_HONEY_BLOCK) || !state.is(BovinesTags.BlockTags.DOES_NOT_STICK_RICH_HONEY_BLOCK) && state2.is(BovinesBlocks.RICH_HONEY_BLOCK))
             return false;
         return original;
     }