+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/mixin/ServerAdvancementManagerMixin.java
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 @Mixin(ServerAdvancementManager.class)
 public class ServerAdvancementManagerMixin {
-    @ModifyVariable(method = { "lambda$apply$0", "method_20723"}, at = @At(value = "HEAD"), argsOnly = true)
+    @ModifyVariable(method = { "lambda$apply$0", "method_20723" }, at = @At(value = "HEAD"), argsOnly = true)
     private JsonElement bovinesandbuttercups$applyAdvancement(JsonElement element, @Local(argsOnly = true) ResourceLocation id) {
         if (element.isJsonObject() && element.getAsJsonObject().getAsJsonPrimitive("bovinesandbuttercups:pack_source").getAsString().equals("vanilla")) {
             if (id.equals(ResourceLocation.withDefaultNamespace("husbandry/bred_all_animals")))