+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/util/MooshroomChildTypeUtil.java
@@ -6,7 +6,6 @@ import house.greenhouse.bovinesandbuttercups.api.CowType;
 import house.greenhouse.bovinesandbuttercups.api.attachment.CowTypeAttachment;
 import house.greenhouse.bovinesandbuttercups.api.cowtype.OffspringConditions;
 import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.BreedCowWithTypeTrigger;
-import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.MutationTrigger;
 import house.greenhouse.bovinesandbuttercups.content.data.configuration.MooshroomConfiguration;
 import house.greenhouse.bovinesandbuttercups.registry.BovinesLootContextParamSets;
 import house.greenhouse.bovinesandbuttercups.registry.BovinesLootContextParams;
@@ -60,8 +59,7 @@ public class MooshroomChildTypeUtil {
             createParticles(child, randomType, parent.position());
 
             if (parent.getLoveCause() != null)
-                MutationTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, (Holder) randomType);
-            BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, (Holder) randomType);
+                BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, true, (Holder) randomType);
             return randomType;
         }
 
@@ -73,11 +71,13 @@ public class MooshroomChildTypeUtil {
             return null;
 
         if (!otherType.equals(parentType) && parent.getRandom().nextBoolean()) {
-            BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, (Holder<CowType<?>>)(Holder<?>)otherType);
+            if (parent.getLoveCause() != null)
+                BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, false, (Holder<CowType<?>>)(Holder<?>)otherType);
             return otherType;
         }
 
-        BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, (Holder<CowType<?>>)(Holder<?>)parentType);
+        if (parent.getLoveCause() != null)
+            BreedCowWithTypeTrigger.INSTANCE.trigger(parent.getLoveCause(), parent, other, child, false, (Holder<CowType<?>>)(Holder<?>)parentType);
         return parentType;
     }